Define Google Earth Engine data to collect

Define Google Earth Engine data to collect

Usage

collect_ee(
  collection = NULL,
  coords = NULL,
  date = NULL,
  end_date = NULL,
  buffer = NULL,
  bound = FALSE,
  config = NULL
)

Arguments

collection

string: name of a Google Earth Engine collection. Collections can be found on the Google Earth Engine Catalog

coords

numeric: GPS coordinates in WGS84 [East, North]. Minimum of one set of coordinates should be provided to create a point coordinate. If more than one set of coordinates is provided, a polygon will be created

date

string: Start date of image(s) to be collected in YYYY-MM-DD or YYYY format. If YYY-MM-DD is provided, will search the specific date only (which may not contain an image), unless end_date is also provided, which will collect images between the two dates. If YYYY is provided, all images within the specified year will be included in the collection

end_date

string, optional: When paired with date argument, can define a date range in YYYY-MM-DD or YYYY format

buffer

integer, optional: If coords is a single point, buffer can be used to create a circular buffer with the specific radius in metres. If coords contains more than one set of coordinates, this argument does nothing

bound

logical, optional: If buffer contains an integer value, this agrument will convert the circular buffer to a bounding box instead. Defaults to FALSE

config

string, optional: Path to a configuration file in .yaml format. When this is provided, all arguments are ignored and the function will refer to the configuration file to determine argument values

Value

an object containing attributes necessary to preprocess and and download images for all other *_ee() functions

Examples

NULL
#> NULL

collect_ee(
  collection = "LANDSAT/LC09/C02/T1_L2",
  coords = c(149.769345, -30.335861, 149.949173, -30.206271),
  date = "2021-06-01",
  end_date = "2022-06-01"
)
#> Error in py_module_import(module, convert = convert): ModuleNotFoundError: No module named 'dateutil'