Download from SLGA (Soil Atributes)

Wrapper funtion to get layers from the Soil and Landscape Grid of Australia (SLGA).

Usage

download_slga(
  layer,
  bounding_box,
  out_path,
  resolution = 3,
  depth_min = 0,
  depth_max = 200,
  get_ci = TRUE,
  verbose = FALSE
)

Arguments

layer

Layer name(s) to download as a vector of strings

bounding_box

A numeric vector of coordinate values used to generate a bounding box around are of interest in EPSG:4326 format, for example, c(min_x, min_y, max_x, max_y)

out_path

Path to an output/download folder, e.g. 'downloads/'

resolution

An integer value for resolution, in arc seconds

depth_min, depth_max

SLGA layers can be filtered between specific depths. Minimum and maximum depths can be set using depth_min and depth_max. Defaults to 0 and 200, respectively

get_ci

Also download upper and lower 95% confidence limits with the the layer. Defaults to TRUE

verbose

Print as much information as possible to the debug log

Value

a list of filenames (after files have been downloaded or processed)