Visualise an Earth Engine Image or Image Collection on an interactive map

A folium map is produced and image(s) collected so far are displayed as layer(s) on top of the map.

Usage

map_ee(object, bands = NULL, minmax = NULL, palette = NULL)

Arguments

object

object: a data object produced by collect_e()

bands

string, optional: a string or list of strings representing the bands to be visualised. If NULL, will present a list of available bands to visualise

minmax

numeric, optional: A list of two integers representing the minimum and maximum values for image pixel colours in single-band images. If set to NULL, the min and max values are automatically calculated Defaults to NULL

palette

string, optional: A string representing the name of a palette to be used for map colors. Names are accessed from Matplotlib Colourmaps as described in https://matplotlib.org/stable/tutorials/colors/colormaps.html. In addition, "ndvi", "ndwi" and "terrain" palettes are available. If set to None, "viridis" is used. Defaults to NULL

Value

an updated collect_ee() object that can be passed on to aggregate_ee() or download_ee()

Examples

NULL
#> NULL